V 2.13
Found a bug in parallel node calculation for some inputs, will work on
it after class. Also developing a system to copy input components/nodes
so that the original list is preserved. This will both allow the user to
continue to enter components after a calculation, and allow node
specific information to be calculated.