V 2.302
Moved resistor node sorting to circuitanalysis.java, in order to sort
only the copy objects such that the user list that they can continue to
work on after calculations will maintain the node ordering that the user
input. This is a nice to have feature because I found it easier to think
of the last resistor as Node n to Node 0, so by only sorting the copies
that last resistor will maintain its order for the user.