V 2.403
Compartmentalized the main method, taking code out of the main and
implementing it as methods that the main method calls. While this may
add a small amount of overhead, it will make the main method much more
readable, scalable and maintainable as per industry best practices.
Would also help the main method be worked on in a team environment.